+++ 01-Authorization-RS256/middleware/jwt.go
@@ -1,0 +1,77 @@
+package middleware
+
+import (
+	"encoding/json"
+	"errors"
+	"net/http"
+	"os"
+
+	"github.com/auth0/go-jwt-middleware"
+	"github.com/form3tech-oss/jwt-go"
+)
+
+var JWT = jwtmiddleware.New(jwtmiddleware.Options{
+	ValidationKeyGetter: func(token *jwt.Token) (interface{}, error) {
+		// Verify 'aud' claim
+		aud := os.Getenv("AUTH0_AUDIENCE")
+		checkAud := token.Claims.(jwt.MapClaims).VerifyAudience(aud, false)
+		if !checkAud {
+			return token, errors.New("invalid audience")
+		}
+
+		// Verify 'iss' claim
+		iss := "https://" + os.Getenv("AUTH0_DOMAIN") + "/"
+		checkIss := token.Claims.(jwt.MapClaims).VerifyIssuer(iss, false)
+		if !checkIss {
+			return token, errors.New("invalid issuer")
+		}
+
+		cert, err := getPemCert(token)
+		if err != nil {
+			return token, err
+		}
+
+		return jwt.ParseRSAPublicKeyFromPEM([]byte(cert))
+	},
+	SigningMethod: jwt.SigningMethodRS256,
+})
+
+type Jwks struct {
+	Keys []JSONWebKeys `json:"keys"`
+}
+
+type JSONWebKeys struct {
+	Kty string   `json:"kty"`
+	Kid string   `json:"kid"`
+	Use string   `json:"use"`
+	N   string   `json:"n"`
+	E   string   `json:"e"`
+	X5c []string `json:"x5c"`
+}
+
+func getPemCert(token *jwt.Token) (string, error) {
+	resp, err := http.Get("https://" + os.Getenv("AUTH0_DOMAIN") + "/.well-known/jwks.json")
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+
+	var jwks Jwks
+	if err = json.NewDecoder(resp.Body).Decode(&jwks); err != nil {
+		return "", err
+	}
+
+	var cert string
+	for _, key := range jwks.Keys {
+		if token.Header["kid"] == key.Kid {
+			cert = "-----BEGIN CERTIFICATE-----\n" + key.X5c[0] + "\n-----END CERTIFICATE-----"
+			break
+		}
+	}
+
+	if cert == "" {
+		return cert, errors.New("unable to find appropriate key")
+	}
+
+	return cert, nil
+}
